+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This is a backwards breaking change in the crate.
   streams of bytes.
 * The Engine and its Error have been moved into a `rpick::engine` module.
 * The configuration structs, enums, and functions have been moved into a `rpick::config` module.
+* The Engine `ValueError` struct was replaced by a new `PickError` enum.
+
 
 # 0.7.2